requirements.txt
playwright==1.44.0
google-generativeai==0.7.2
python-docx==1.1.0
python-dotenv==1.0.0
requests==2.31.0
click==8.1.7
colorama==0.4.6
tqdm==4.66.1 